Create Testimonial,Call to Action & Footer Section
Create a testimonial section to showcase what our customers have to say
about the product.
Addded a quote from Mike of Burger and Relish and added an image of the
B&R logo in the img folder
Create a call to action section where any catering businesses or
restaurants can input their email addresses so they are kept in the loop
on the progress of our product. They are also notified of any new
developments and when we launch.
Pending integration of mailchimp with the form.
Create a footer section highlighting the various components of our
homepage and with our social media handles.
Lorem ipsum for the SEO section.
Refine CSS styling and assign appropriate HTML tags