Rename Categories on the navbar
Change the categories on the navbar from Home, How it works, Pricing, About us,
Contact us to Home, Benefits, Features and Sign up.
Modify list style type of the lists in the Benefits section to 'circle'
Modify description in the benefits section.
Delete section li color so the purple color is consistent for all h2
tags